requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.12.30
instagrapi==2.0.0
python-dotenv==1.0.0
Pillow>=8.1.1